chore: relocate tests under src
